requirements.txt
csv
re
pytesseract
easyocr
azure-mgmt-resource
azure-ai-formrecognizer
langchain_community
ollama
spacy